Simplify provider presentation flow
