+++ apps/desktop-tauri/src/App.test.tsx
@@ -94,6 +94,7 @@ function settings(overrides: Partial<SettingsSnapshot> = {}): SettingsSnapshot {
     trayScalePercent: 100,
     powertoysStatusPipeEnabled: false,
     claudeAvoidKeychainPrompts: false,
+    codexSparkUsageVisible: true,
     disableKeychainAccess: false,
     providerMetrics: {},
     floatBarEnabled: false,
